JGroups: Switch to "per-destination" bundler for jdbc-ping
Closes #39545
