+++ model/infinispan/src/main/java/org/keycloak/infinispan/util/InfinispanUtils.java
@@ -31,6 +31,10 @@ public final class InfinispanUtils {
     private InfinispanUtils() {
     }
 
+    private static final String INFINISPAN_VIRTUAL_THREADS_PROP = "org.infinispan.threads.virtual";
+
+    private static final int MIN_VT_POOL_SIZE = 2;
+
     // all providers have the same order
     public static final int PROVIDER_ORDER = 1;
 
@@ -92,5 +96,34 @@ public final class InfinispanUtils {
 
     public static void retryBaseTimeMillisToOperationalInfo(Map<String, String> map, int value) {
         map.put(CONFIG_RETRIES_BASE_TIME_MILLIS, Integer.toString(value));
+    }
+
+    public static boolean isVirtualThreadsEnabled() {
+        return Boolean.parseBoolean(System.getProperty(INFINISPAN_VIRTUAL_THREADS_PROP));
+    }
+
+    public static void configureVirtualThreads() {
+        // enable Infinispan and JGroups virtual threads by default
+        if (System.getProperty(INFINISPAN_VIRTUAL_THREADS_PROP) == null && getParallelism() >= MIN_VT_POOL_SIZE)
+            System.setProperty(INFINISPAN_VIRTUAL_THREADS_PROP, "true");
+    }
+
+    public static void ensureVirtualThreadsParallelism() {
+        if (isVirtualThreadsEnabled()) {
+            if (getParallelism() < MIN_VT_POOL_SIZE) {
+                throw new RuntimeException("To be able to use Infinispan/JGroups virtual threads, you need to set the Java system property jdk.virtualThreadScheduler.parallelism to at least " + MIN_VT_POOL_SIZE);
+            }
+        }
+    }
+
+    private static int getParallelism() {
+        int parallelism;
+        String parallelismValue = System.getProperty("jdk.virtualThreadScheduler.parallelism");
+        if (parallelismValue != null) {
+            parallelism = Integer.parseInt(parallelismValue);
+        } else {
+            parallelism = Runtime.getRuntime().availableProcessors();
+        }
+        return parallelism;
     }
 }
